Add solution for finding the minimum operations
This commit adds a solution for finding the minimum operations required to reach a given number 'n'. The code implements a recursive function 'minOperations' that calculates the minimum number of operations needed to reach 'n' by either doubling the current value or incrementing it by 1. The function returns the minimum number of operations required.